created ConsoleMessage component to handle all calls to console
